+++ app/src/main/java/ayds/lisboa/songinfo/moredetails/home/model/MoreDetailsModelInjector.kt
@@ -1,0 +1,2 @@
+package ayds.lisboa.songinfo.moredetails.home.model
+
