+++ app/src/main/java/ayds/lisboa/songinfo/moredetails/home/model/repository/external/lastfm/LastFMAPI.java
@@ -1,4 +1,4 @@
-package ayds.lisboa.songinfo.moredetails.home.model;
+package ayds.lisboa.songinfo.moredetails.home.model.repository.external.lastfm;
 
 
 import retrofit2.Call;
